LPS-65453 Extract logic to abstract base class without changes in order to be able reuse it in other workflowContext upgrades
